Deleted and recreated SearchLoader
